+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.7.8 (WIP)
 
 * Add some Terser workarounds
+* Make `String#replaceAll()` a polyfill, as it is now part of ECMAScript-2021
 
 ## 0.7.7 (2021-01-21)